+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ConstrainSolver.jl - Changelog
  
+## Unreleased
+- Support for indicator/reified in indicator/reified (without bridges) [PR #251](https://github.com/Wikunia/ConstraintSolver.jl/pull/251)
+  
 ## v0.6.9 (17th of July 2021)
 - set activator to false when inner violated [PR #266](https://github.com/Wikunia/ConstraintSolver.jl/pull/266)